Move PoliciesCommand tests into own js module
