+++ gsa/src/gmp/commands/__tests__/policies.js
@@ -1,0 +1,80 @@
+/* Copyright (C) 2019 Greenbone Networks GmbH
+ *
+ * SPDX-License-Identifier: GPL-2.0-or-later
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 2
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ */
+import {ALL_FILTER} from 'gmp/models/filter';
+
+import {createEntitiesResponse, createHttp} from '../testing';
+
+import {PoliciesCommand} from '../policies';
+
+describe('PoliciesCommand tests', () => {
+  test('should return all policies', () => {
+    const response = createEntitiesResponse('config', [
+      {
+        _id: '1',
+      },
+      {
+        _id: '2',
+      },
+    ]);
+
+    const fakeHttp = createHttp(response);
+
+    expect.hasAssertions();
+
+    const cmd = new PoliciesCommand(fakeHttp);
+    return cmd.getAll().then(resp => {
+      expect(fakeHttp.request).toHaveBeenCalledWith('get', {
+        args: {
+          cmd: 'get_configs',
+          filter: ALL_FILTER.toFilterString(),
+          usage_type: 'policy',
+        },
+      });
+      const {data} = resp;
+      expect(data.length).toEqual(2);
+    });
+  });
+
+  test('should return policies', () => {
+    const response = createEntitiesResponse('config', [
+      {
+        _id: '1',
+      },
+      {
+        _id: '2',
+      },
+    ]);
+
+    const fakeHttp = createHttp(response);
+
+    expect.hasAssertions();
+
+    const cmd = new PoliciesCommand(fakeHttp);
+    return cmd.get().then(resp => {
+      expect(fakeHttp.request).toHaveBeenCalledWith('get', {
+        args: {
+          cmd: 'get_configs',
+          usage_type: 'policy',
+        },
+      });
+      const {data} = resp;
+      expect(data.length).toEqual(2);
+    });
+  });
+});
