+++ gsa/src/gmp/commands/__tests__/policy.js
@@ -16,16 +16,13 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
  */
-import {ALL_FILTER} from 'gmp/models/filter';
-
 import {
   createActionResultResponse,
   createEntityResponse,
-  createEntitiesResponse,
   createHttp,
 } from '../testing';
 
-import {PolicyCommand, PoliciesCommand} from '../policies';
+import {PolicyCommand} from '../policies';
 
 describe('PolicyCommand tests', () => {
   test('should create new policy', () => {
@@ -171,63 +168,6 @@ describe('PolicyCommand tests', () => {
 
       const {data} = resp;
       expect(data.id).toEqual('foo');
-    });
-  });
-});
-
-describe('PoliciesCommand tests', () => {
-  test('should return all policies', () => {
-    const response = createEntitiesResponse('config', [
-      {
-        _id: '1',
-      },
-      {
-        _id: '2',
-      },
-    ]);
-
-    const fakeHttp = createHttp(response);
-
-    expect.hasAssertions();
-
-    const cmd = new PoliciesCommand(fakeHttp);
-    return cmd.getAll().then(resp => {
-      expect(fakeHttp.request).toHaveBeenCalledWith('get', {
-        args: {
-          cmd: 'get_configs',
-          filter: ALL_FILTER.toFilterString(),
-          usage_type: 'policy',
-        },
-      });
-      const {data} = resp;
-      expect(data.length).toEqual(2);
-    });
-  });
-
-  test('should return policies', () => {
-    const response = createEntitiesResponse('config', [
-      {
-        _id: '1',
-      },
-      {
-        _id: '2',
-      },
-    ]);
-
-    const fakeHttp = createHttp(response);
-
-    expect.hasAssertions();
-
-    const cmd = new PoliciesCommand(fakeHttp);
-    return cmd.get().then(resp => {
-      expect(fakeHttp.request).toHaveBeenCalledWith('get', {
-        args: {
-          cmd: 'get_configs',
-          usage_type: 'policy',
-        },
-      });
-      const {data} = resp;
-      expect(data.length).toEqual(2);
     });
   });
 });
